+++ src/restore.js
@@ -1,31 +1,23 @@
 import * as cache from "@actions/cache";
 import * as core from "@actions/core";
-import { CACHE_KEY_PREFIX, deleteCacheEntry, getCacheRestorePath, getExistingCacheEntries } from "./helpers.js";
+import { getExistingCacheEntries } from "./helpers.js";
 
 const token = core.getInput("token");
 const vcpkgArchivePath = core.getInput("archive-path");
 
 await core.group("Restoring vcpkg cache", async () => {
-  const actionsCaches = await getExistingCacheEntries(token);
+  const actionsCaches = await getExistingCacheEntries(token, vcpkgArchivePath);
 
   if (actionsCaches.length < 1) {
-    core.info(`No cache entries found with prefix '${CACHE_KEY_PREFIX}'`);
+    core.info(`No cache entries found with prefix '${vcpkgArchivePath}'`);
     return;
   }
 
   await Promise.all(
     actionsCaches.map(async (cacheKey) => {
-      const archivePath = getCacheRestorePath(vcpkgArchivePath, cacheKey);
-      core.info(`Restoring '${cacheKey}' to '${archivePath}'`);
+      core.info(`Restoring '${cacheKey}'`);
 
-      const savedCacheKey = await cache.restoreCache([archivePath], cacheKey, undefined, undefined, true);
-      if (!savedCacheKey) {
-        core.warning(
-          "Cache failed to restore (likely due to a version mismatch from using a different archive path). Deleting old entry"
-        );
-
-        await deleteCacheEntry(token, cacheKey);
-      }
+      await cache.restoreCache([cacheKey], cacheKey, undefined, undefined, true);
     })
   );
 });