+++ src/save.js
@@ -2,13 +2,13 @@ import * as cache from "@actions/cache";
 import * as core from "@actions/core";
 import * as path from "path";
 import * as fs from "fs/promises";
-import { getCacheKey, getExistingCacheEntries } from "./helpers.js";
+import { getExistingCacheEntries } from "./helpers.js";
 
 const token = core.getInput("token");
 const vcpkgArchivePath = core.getInput("archive-path");
 
 await core.group("Saving vcpkg cache", async () => {
-  const actionsCaches = new Set(await getExistingCacheEntries(token));
+  const actionsCaches = new Set(await getExistingCacheEntries(token, vcpkgArchivePath));
 
   try {
     const directories = await fs.readdir(vcpkgArchivePath, { withFileTypes: true });
@@ -21,7 +21,6 @@ await core.group("Saving vcpkg cache", async () => {
       const subfolderPath = path.join(vcpkgArchivePath, directory.name);
       const files = await fs.readdir(subfolderPath, { withFileTypes: true });
       for (const file of files) {
-        const cacheKey = getCacheKey(file.name);
         const archivePath = path.join(vcpkgArchivePath, directory.name, file.name);
 
         if (!file.isFile() && !file.name.endsWith(".zip")) {
@@ -29,14 +28,14 @@ await core.group("Saving vcpkg cache", async () => {
           continue;
         }
 
-        if (actionsCaches.has(cacheKey)) {
+        if (actionsCaches.has(archivePath)) {
           core.info(`Skipping '${archivePath}' as already present in cache`);
           continue;
         }
 
-        core.info(`Saving '${archivePath}' to '${cacheKey}'`);
+        core.info(`Saving '${archivePath}'`);
 
-        await cache.saveCache([archivePath], getCacheKey(file.name), undefined, true);
+        await cache.saveCache([archivePath], archivePath, undefined, true);
       }
     }
   } catch (error) {